complex_solution/DjangoFastAPIMinio/fastApiProject/requirements.txt
aioboto3==11.2.0
aiobotocore==2.5.0
aiofile==3.8.7
aiohttp==3.8.4
aioitertools==0.11.0
aiosignal==1.3.1
anyio==3.7.0
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
boto3==1.26.76
botocore==1.29.76
caio==0.9.12
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Django==4.2.2
fastapi==0.97.0
frozenlist==1.3.3
h11==0.14.0
idna==3.4
jmespath==1.0.1
loguru==0.7.0
miniopy-async==1.15
multidict==6.0.4
orjson==3.9.1
pydantic==1.10.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
s3transfer==0.6.1
shortuuid==1.0.11
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.16
sqlparse==0.4.4
starlette==0.27.0
typing_extensions==4.6.3
urllib3==1.26.16
uvicorn==0.22.0
wrapt==1.15.0
yarl==1.9.2
